[feat]: 세션필터 구현 및 유형검사 rest api 구현

+++ src/main/java/com/coders/boardgame/domain/user/dto/PlayerDto.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 public class PlayerDto {
     private Long userId;
     private String name;
-    private int score;
+    private byte score;
     private Integer usedTime = 0;
     private int puzzlePiece = 0;
     private boolean isSpeaker;
